Revert "[SPARK-29277][SQL] Add early DSv2 filter and projection pushdown"
This reverts commit cfc80d0eb18e1ec2866204da3500acd5f4dde2ea.
