+++ sql/catalyst/src/main/scala/org/apache/spark/sql/execution/datasources/v2/DataSourceV2Relation.scala
@@ -26,7 +26,6 @@ import org.apache.spark.sql.connector.read.{Scan, ScanBuilder, Statistics => V2S
 import org.apache.spark.sql.connector.read.streaming.{Offset, SparkDataStream}
 import org.apache.spark.sql.connector.write.WriteBuilder
 import org.apache.spark.sql.util.CaseInsensitiveStringMap
-import org.apache.spark.util.Utils
 
 /**
  * A logical plan representing a data source v2 table.
@@ -51,53 +50,12 @@ case class DataSourceV2Relation(
     s"RelationV2${truncatedString(output, "[", ", ", "]", maxFields)} $name"
   }
 
-  override def computeStats(): Statistics = {
-    if (Utils.isTesting) {
-      // when testing, throw an exception if this computeStats method is called because stats should
-      // not be accessed before pushing the projection and filters to create a scan. otherwise, the
-      // stats are not accurate because they are based on a full table scan of all columns.
-      throw new IllegalStateException(
-        s"BUG: computeStats called before pushdown on DSv2 relation: $name")
-    } else {
-      // when not testing, return stats because bad stats are better than failing a query
-      table.asReadable.newScanBuilder(options) match {
-        case r: SupportsReportStatistics =>
-          val statistics = r.estimateStatistics()
-          DataSourceV2Relation.transformV2Stats(statistics, None, conf.defaultSizeInBytes)
-        case _ =>
-          Statistics(sizeInBytes = conf.defaultSizeInBytes)
-      }
-    }
-  }
-
-  override def newInstance(): DataSourceV2Relation = {
-    copy(output = output.map(_.newInstance()))
-  }
-}
-
-/**
- * A logical plan for a DSv2 table with a scan already created.
- *
- * This is used in the optimizer to push filters and projection down before conversion to physical
- * plan. This ensures that the stats that are used by the optimizer account for the filters and
- * projection that will be pushed down.
- *
- * @param table a DSv2 [[Table]]
- * @param scan a DSv2 [[Scan]]
- * @param output the output attributes of this relation
- */
-case class DataSourceV2ScanRelation(
-    table: Table,
-    scan: Scan,
-    output: Seq[AttributeReference]) extends LeafNode with NamedRelation {
-
-  override def name: String = table.name()
-
-  override def simpleString(maxFields: Int): String = {
-    s"RelationV2${truncatedString(output, "[", ", ", "]", maxFields)} $name"
+  def newScanBuilder(): ScanBuilder = {
+    table.asReadable.newScanBuilder(options)
   }
 
   override def computeStats(): Statistics = {
+    val scan = newScanBuilder().build()
     scan match {
       case r: SupportsReportStatistics =>
         val statistics = r.estimateStatistics()
@@ -105,6 +63,10 @@ case class DataSourceV2ScanRelation(
       case _ =>
         Statistics(sizeInBytes = conf.defaultSizeInBytes)
     }
+  }
+
+  override def newInstance(): DataSourceV2Relation = {
+    copy(output = output.map(_.newInstance()))
   }
 }
